around the ft sections
